Move some test functions from tests to src
This allows them to be used for unit tests
